(refactor) added an endpoint that pulls patients who are marked as deceased
