--- packages/esm-morgue-app/src/card/row-elements/row-elements.scss
@@ -1,16 +1,0 @@
-@use '@carbon/styles/scss/spacing';
-@use '@carbon/styles/scss/type';
-@import '~@openmrs/esm-styleguide/src/vars';
-
-.unitNumber {
-  @include type.type-style('heading-compact-01');
-  border-radius: 50%;
-  background-color: $color-blue-10;
-  color: $color-blue-60-2;
-  padding: spacing.$spacing-04;
-  width: spacing.$spacing-03;
-  height: spacing.$spacing-03;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
