--- packages/esm-morgue-app/src/card/row-elements/unit-patient-age.tsx
@@ -1,9 +1,0 @@
-import React from 'react';
-import { useTranslation } from 'react-i18next';
-import { age } from '@openmrs/esm-framework';
-
-const UnitPatientAge = () => {
-  return <div>{age('12')}</div>;
-};
-
-export default UnitPatientAge;
