--- packages/esm-morgue-app/src/card/row-elements/unit-patient-unit-number.tsx
@@ -1,12 +1,0 @@
-import React from 'react';
-import styles from '../unit-patient-card.scss';
-
-const UnitPatientUnitNumber = () => {
-  return (
-    <div className={styles.UnitNumberBox}>
-      <span className={styles.unitPatientUnitNumber}>{'1'}</span>
-    </div>
-  );
-};
-
-export default UnitPatientUnitNumber;
