--- packages/esm-morgue-app/src/card/unit-patient-card.scss
@@ -1,129 +1,0 @@
-@use '@carbon/styles/scss/spacing';
-@use '@carbon/styles/scss/type';
-@use '@carbon/colors';
-@import '~@openmrs/esm-styleguide/src/vars';
-
-.wardPatientCard {
-  @include type.type-style('body-compact-01');
-  color: $text-02;
-
-  display: flex;
-  flex-wrap: wrap;
-  align-items: center;
-  gap: spacing.$spacing-02;
-  background-color: $ui-02;
-
-  position: relative; // this allows positioning the button correctly
-
-  > .wardPatientCardRow:not(:first-child) {
-    border-top: 1px colors.$gray-20 solid;
-  }
-}
-
-.wardPatientCardButton {
-  border: none;
-  padding: 0;
-
-  &::before {
-    content: '';
-    position: absolute;
-    inset: 0;
-    z-index: 1;
-    cursor: pointer;
-    border: 2px solid transparent;
-    transition: border-color 200ms;
-  }
-
-  &:hover::before,
-  &:focus::before {
-    border-color: $interactive-01;
-  }
-
-  &:focus {
-    outline: none;
-  }
-}
-
-.activeWardPatientCardButton {
-  &::before {
-    content: '';
-    position: absolute;
-    inset: 0;
-    z-index: 1;
-    cursor: pointer;
-    border: 2px solid $interactive-01;
-    transition: border-color 200ms;
-  }
-}
-
-.wardPatientCardRow {
-  width: 100%;
-  padding: spacing.$spacing-04;
-}
-
-.wardPatientCardRow:empty {
-  display: none;
-}
-
-.wardPatientCardHeader {
-  @extend .dotSeparatedChildren;
-  display: flex;
-  flex-wrap: wrap;
-}
-
-.unitPatientName {
-  @include type.type-style('heading-compact-02');
-  color: $text-02;
-
-  &::before {
-    content: '' !important;
-  }
-}
-
-.unitPatientUnitNumber {
-  @include type.type-style('heading-compact-01');
-  border-radius: 50%;
-  color: $ui-02;
-  background-color: $color-blue-60-2;
-  padding: spacing.$spacing-04;
-  width: spacing.$spacing-04;
-  height: spacing.$spacing-04;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-
-  &.empty {
-    background-color: $color-blue-10;
-    color: $color-blue-60-2;
-  }
-}
-
-.UnitNumberBox {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.unitPatientAddress {
-  @extend .dotSeparatedChildren;
-  display: flex;
-  flex-wrap: wrap;
-  align-items: center;
-  gap: spacing.$spacing-02;
-}
-
-.UnitPatientObsLabel {
-  padding-right: spacing.$spacing-02;
-}
-
-.dotSeparatedChildren {
-  > div:not(div:first-of-type) {
-    display: flex;
-    align-items: center;
-
-    &::before {
-      content: '·';
-      padding: 0 spacing.$spacing-02;
-    }
-  }
-}
