+++ packages/esm-morgue-app/src/hook/useMorgue.resource.ts
@@ -62,10 +62,11 @@ export function useCauseOfDeathConcept() {
   return result;
 }
 
-export const useDeceasedPatient = (deceasedPatientName: String) => {
+export const useDeceasedPatient = () => {
   const customRepresentation =
     'custom:(uuid,display,identifiers:(identifier,uuid,preferred,location:(uuid,name)),person:(uuid,display,gender,birthdate,dead,age,deathDate,causeOfDeath:(uuid,display),preferredAddress:(uuid,stateProvince,countyDistrict,address4)))';
-  const url = `/ws/rest/v1/patient?v=${customRepresentation}&dead=true&q=${deceasedPatientName}`;
+  const url = `/ws/rest/v1//morgue/patient?v=${customRepresentation}&dead=true`;
+
   const { data, error, isLoading, mutate } = useSWRImmutable<{ data: DeceasedPatientResponse }>(url, openmrsFetch);
 
   const deceasedPatient = data?.data?.results || null;
