+++ packages/esm-morgue-app/src/types/index.ts
@@ -101,3 +101,38 @@ export interface Observation {
   value: any;
   obsDatetime?: string;
 }
+export interface DeceasedInfo {
+  uuid: string;
+  display: string;
+  identifiers: Array<{
+    identifier: string;
+    uuid: string;
+    preferred: boolean;
+    location: {
+      uuid: string;
+      name: string;
+    };
+  }>;
+  person: {
+    uuid: string;
+    display: string;
+    gender: string;
+    birthdate: string;
+    dead: boolean;
+    age: number;
+    deathDate: string | null;
+    causeOfDeath: {
+      uuid: string;
+      display: string;
+    } | null;
+    preferredAddress: {
+      uuid: string;
+      stateProvince: string | null;
+      countyDistrict: string | null;
+      address4: string | null;
+    } | null;
+  };
+}
+export interface DeceasedPatientResponse {
+  results: DeceasedInfo[];
+}
